+++ app/services/get_contentful_question.rb
@@ -1,8 +1,28 @@
 require "contentful"
 
 class GetContentfulQuestion
+  class EntryNotFound < StandardError
+    attr_accessor :message
+    def initialize(message)
+      @message = message
+    end
+  end
+
   def call(entry_id:)
-    contentful_client.entry(entry_id).raw
+    response = contentful_client.entry(entry_id)
+    if response.nil?
+      error_message = "The following Contentful entry identifier could not be found."
+      Rollbar.warning(
+        error_message,
+        contentful_url: ENV["CONTENTFUL_URL"],
+        contentful_space_id: ENV["CONTENTFUL_SPACE"],
+        contentful_environment: ENV["CONTENTFUL_ENVIRONMENT"],
+        contentful_entry_id: entry_id
+      )
+      raise EntryNotFound.new(error_message)
+    end
+
+    response.raw
   end
 
   private